Fix signout redirect & locales issue with en-UK
