+++ src/components/header/HeaderDrawer.js
@@ -33,7 +33,7 @@ class HeaderDrawer extends Component {
     e.preventDefault();
     this.props.handleDrawerClose();
     this.props.authStore.logout().then(() => {
-      window.location.href = UrlService.createUrl(window.location.host, '/' + undefsafe(this.props.orgStore.currentOrganisation, 'tag'), null);
+      window.location.href = UrlService.createUrl(window.location.host, this.props.orgStore.currentOrganisation ?  '/' + this.props.orgStore.currentOrganisation.tag : '' , null);
     });
   }
